+++ src/Force.hpp
@@ -12,15 +12,15 @@ struct Force
 
 struct Viscosity : Force
 {
-    float _viscosity;
-    Viscosity(float viscosity);
+    double _viscosity;
+    Viscosity(double viscosity);
     vec3 computeForce(Object & object);
 };
 
 struct Gravity : Force
 {
     vec3 _gravity;
-    Gravity(float x, float y, float z);
+    Gravity(double x, double y, double z);
     vec3 computeForce(Object & object);
 };
